requirements-unified.txt
# Unified requirements for workspace tasks and CI convenience

# Core runtime
fastapi>=0.110.0
uvicorn>=0.23.0
requests>=2.31.0
httpx>=0.27.0
pandas>=2.0.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
python-dotenv>=1.0.0
tomli; python_version < "3.11"

# Optional platform bits
pywin32>=306; sys_platform == "win32"

# Scheduling
apscheduler>=3.10.4

# Testing & quality
pytest>=7.4.0
pytest-asyncio>=0.21.0
hypothesis>=6.98.0
pytest-cov>=4.1.0
ruff>=0.4.0
mypy>=1.10.0
pyright>=1.1.373
pydoclint>=0.4.1
interrogate>=1.7.0
black>=24.3.0
flake8>=7.0.0
pylint>=3.2.0
